Update to account for EndlessRoadCar being moved to dev/.
