+++ drake/automotive/automotive_simulator.h
@@ -7,8 +7,8 @@
 #include <vector>
 
 #include "drake/automotive/curve2.h"
+#include "drake/automotive/dev/endless_road_car.h"
 #include "drake/automotive/endless_road_car_to_euler_floating_joint.h"
-#include "drake/automotive/endless_road_car.h"
 #include "drake/automotive/maliput/api/road_geometry.h"
 #include "drake/automotive/maliput/utility/infinite_circuit_road.h"
 #include "drake/automotive/simple_car.h"
